# sistema/src/main/java/com/sistemajuridico/backend/presentation/controllers/TarefaController.java
package com.sistemajuridico.backend.presentation.controllers;

import com.sistemajuridico.backend.presentation.openapi.TarefaControllerOpenApi;

import com.sistemajuridico.backend.core.domain.Tarefa;
import com.sistemajuridico.backend.core.domain.Usuario;
import com.sistemajuridico.backend.core.domain.enums.TipoTarefaEnum;
import com.sistemajuridico.backend.core.domain.exceptions.RecursoNaoEncontradoException;
import com.sistemajuridico.backend.core.domain.exceptions.RegraNegocioException;
import com.sistemajuridico.backend.core.usecases.*;
import com.sistemajuridico.backend.infrastructure.persistence.UsuarioRepository;
import com.sistemajuridico.backend.presentation.dtos.TarefaDTO;
import jakarta.validation.Valid;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.security.Principal;
import java.time.LocalDate;
import java.util.ArrayList;
import java.util.List;
import java.util.Optional;
import java.util.UUID;

@RestController
@RequestMapping("/api/tarefas")
public class TarefaController implements TarefaControllerOpenApi {

    private final CriarTarefaUseCase criarTarefaUseCase;
    private final ConcluirTarefaUseCase concluirTarefaUseCase;
    private final AtualizarTarefaUseCase atualizarTarefaUseCase;
    private final ExcluirTarefaUseCase excluirTarefaUseCase;
    private final ListarTarefasDashboardUseCase listarTarefasDashboardUseCase;
    private final ListarTarefasPorPeriodoUseCase listarTarefasPorPeriodoUseCase;
    private final UsuarioRepository usuarioRepository;

    public TarefaController(CriarTarefaUseCase criarTarefaUseCase,
                             ConcluirTarefaUseCase concluirTarefaUseCase,
                             AtualizarTarefaUseCase atualizarTarefaUseCase,
                             ExcluirTarefaUseCase excluirTarefaUseCase,
                             ListarTarefasDashboardUseCase listarTarefasDashboardUseCase,
                             ListarTarefasPorPeriodoUseCase listarTarefasPorPeriodoUseCase,
                             UsuarioRepository usuarioRepository) {
        this.criarTarefaUseCase = criarTarefaUseCase;
        this.concluirTarefaUseCase = concluirTarefaUseCase;
        this.atualizarTarefaUseCase = atualizarTarefaUseCase;
        this.excluirTarefaUseCase = excluirTarefaUseCase;
        this.listarTarefasDashboardUseCase = listarTarefasDashboardUseCase;
        this.listarTarefasPorPeriodoUseCase = listarTarefasPorPeriodoUseCase;
        this.usuarioRepository = usuarioRepository;
    }

    @Override
    @PostMapping
    public ResponseEntity<TarefaDTO> criar(@RequestBody @Valid TarefaDTO dto) {
        Tarefa tarefa = dto.toEntity();
        Tarefa tarefaSalva = this.criarTarefaUseCase.executar(tarefa, dto.usuarioId(), dto.processoId());
        return ResponseEntity.status(HttpStatus.CREATED).body(TarefaDTO.fromEntity(tarefaSalva));
    }

    @Override
    @PutMapping("/{id}")
    public ResponseEntity<TarefaDTO> atualizar(@PathVariable UUID id, @RequestBody @Valid TarefaDTO dto) {
        Tarefa tarefa = dto.toEntity();
        Tarefa tarefaAtualizada = this.atualizarTarefaUseCase.executar(id, tarefa);
        return ResponseEntity.ok(TarefaDTO.fromEntity(tarefaAtualizada));
    }

    @Override
    @DeleteMapping("/{id}")
    public ResponseEntity<Void> excluir(@PathVariable UUID id) {
        this.excluirTarefaUseCase.executar(id);
        return ResponseEntity.noContent().build();
    }

    @Override
    @PatchMapping("/{id}/concluir")
    public ResponseEntity<TarefaDTO> concluir(@PathVariable UUID id) {
        Tarefa tarefaConcluida = this.concluirTarefaUseCase.executar(id);
        return ResponseEntity.ok(TarefaDTO.fromEntity(tarefaConcluida));
    }

    @Override
    @GetMapping("/dashboard/{usuarioId}")
    public ResponseEntity<List<TarefaDTO>> listarDashboard(@PathVariable UUID usuarioId) {
        List<Tarefa> tarefas = this.listarTarefasDashboardUseCase.executar(usuarioId);
        List<TarefaDTO> response = new ArrayList<>();
        for (Tarefa tarefa : tarefas) {
            response.add(TarefaDTO.fromEntity(tarefa));
        }
        return ResponseEntity.ok(response);
    }

    @Override
    @GetMapping("/agenda")
    public ResponseEntity<List<TarefaDTO>> listarAgenda(
            @RequestParam(name = "inicio", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate inicio,
            @RequestParam(name = "fim", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate fim,
            @RequestParam(name = "concluida", required = false) Boolean concluida,
            @RequestParam(name = "status", required = false) String status,
            @RequestParam(name = "tipo", required = false) TipoTarefaEnum tipo,
            @RequestParam(name = "processoId", required = false) UUID processoId,
            @RequestParam(name = "responsavelId", required = false) UUID responsavelId,
            Principal principal) {
        if (principal == null || principal.getName() == null) {
            throw new RegraNegocioException("Usuário não autenticado no sistema!");
        }

        String email = principal.getName();
        Optional<Usuario> optUsuario = this.usuarioRepository.findByEmail(email);
        if (optUsuario.isEmpty()) {
            throw new RecursoNaoEncontradoException("Usuário autenticado não encontrado no sistema!");
        }

        Boolean filtroConcluida = concluida;
        if (filtroConcluida == null && status != null) {
            if ("CONCLUIDA".equalsIgnoreCase(status) || "CONCLUIDO".equalsIgnoreCase(status)) {
                filtroConcluida = Boolean.TRUE;
            } else if ("PENDENTE".equalsIgnoreCase(status)) {
                filtroConcluida = Boolean.FALSE;
            }
        }

        Usuario usuarioLogado = optUsuario.get();
        List<TarefaDTO> response = this.listarTarefasPorPeriodoUseCase.executar(
                usuarioLogado, inicio, fim, filtroConcluida, tipo, processoId, responsavelId
        );
        return ResponseEntity.ok(response);
    }
}