# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sistema Jurídico API",
    "description": "Documentação oficial da API do LegalTech",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/tarefas/{id}": {
      "put": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Atualizar tarefa",
        "description": "Atualiza os dados de uma tarefa existente",
        "operationId": "atualizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TarefaDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Tarefa não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de integridade",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "400": {
            "description": "Dados de requisição inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroValidacaoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "200": {
            "description": "Tarefa atualizada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TarefaDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Excluir tarefa",
        "description": "Remove uma tarefa do sistema",
        "operationId": "excluir",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tarefa excluída com sucesso"
          },
          "409": {
            "description": "Conflito de integridade referencial",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "404": {
            "description": "Tarefa não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/processos/{id}": {
      "get": {
        "tags": [
          "Processos"
        ],
        "summary": "Buscar processo por ID",
        "description": "Recupera os detalhes completos de um processo através do seu identificador único",
        "operationId": "buscarPorId_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Processo não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "200": {
            "description": "Processo encontrado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Processos"
        ],
        "summary": "Atualizar processo",
        "description": "Atualiza os dados de um processo judicial existente",
        "operationId": "atualizar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Processo não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de integridade",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "200": {
            "description": "Processo atualizado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "400": {
            "description": "Dados de requisição inválidos ou campos incorretos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroValidacaoDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/clientes/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Buscar cliente por ID",
        "description": "Recupera os detalhes de um cliente através do seu identificador único",
        "operationId": "buscarPorId_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clientes"
        ],
        "summary": "Atualizar dados do cliente",
        "description": "Atualiza as informações cadastrais de um cliente existente",
        "operationId": "atualizar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Dados de requisição inválidos ou campos não atendem as validações",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroValidacaoDTO"
                }
              }
            }
          },
          "200": {
            "description": "Cliente atualizado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de integridade",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/audiencias/{id}": {
      "get": {
        "tags": [
          "Audiências"
        ],
        "summary": "Buscar audiência por ID",
        "description": "Recupera os detalhes de uma audiência pelo identificador único",
        "operationId": "buscarPorId_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Audiência não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "200": {
            "description": "Audiência retornada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AudienciaDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Audiências"
        ],
        "summary": "Atualizar audiência",
        "description": "Atualiza os dados de uma audiência agendada",
        "operationId": "atualizar_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudienciaDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Audiência não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "200": {
            "description": "Audiência atualizada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AudienciaDTO"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de integridade",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "400": {
            "description": "Dados de requisição inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroValidacaoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Audiências"
        ],
        "summary": "Excluir audiência",
        "description": "Remove um agendamento de audiência",
        "operationId": "excluir_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Audiência não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "204": {
            "description": "Audiência excluída com sucesso"
          },
          "409": {
            "description": "Conflito de integridade referencial",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tarefas": {
      "post": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Criar nova tarefa",
        "description": "Cadastra uma nova tarefa atribuída a um responsável e opcionalmente vinculada a um processo",
        "operationId": "criar_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TarefaDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Usuário responsável ou processo informado não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de integridade",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "400": {
            "description": "Dados de requisição inválidos ou campos obrigatórios ausentes",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroValidacaoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "201": {
            "description": "Tarefa criada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TarefaDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tarefas/{id}/concluir": {
      "patch": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Concluir tarefa",
        "description": "Marca a tarefa especificada como concluída",
        "operationId": "concluir",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa marcada como concluída com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TarefaDTO"
                }
              }
            }
          },
          "404": {
            "description": "Tarefa não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tarefas/dashboard/{usuarioId}": {
      "get": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Listar tarefas para o dashboard",
        "description": "Recupera tarefas prioritárias pendentes para o painel do usuário informado",
        "operationId": "listarDashboard",
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefas do dashboard recuperadas com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TarefaDTO"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboard/{usuarioId}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Obter dashboard por ID de usuário",
        "description": "Recupera as métricas e indicadores de um usuário/advogado específico",
        "operationId": "obterDashboard",
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard recuperado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoDashboardDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProcessoResumoDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "numeroCnj": {
            "type": "string"
          },
          "nomeCliente": {
            "type": "string"
          }
        }
      },
      "TarefaDTO": {
        "required": [
          "dataVencimento",
          "descricao",
          "usuarioId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "descricao": {
            "type": "string"
          },
          "dataVencimento": {
            "type": "string",
            "format": "date"
          },
          "concluida": {
            "type": "boolean"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "DILIGENCIA",
              "PRAZO",
              "CONTATO"
            ]
          },
          "usuarioId": {
            "type": "string",
            "format": "uuid"
          },
          "processo": {
            "$ref": "#/components/schemas/ProcessoResumoDTO"
          }
        }
      },
      "ErroPadraoDTO": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Data e hora em que o erro ocorreu",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP do erro",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Identificador ou classificação do erro",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "description": "Descrição detalhada da mensagem de erro",
            "example": "Dados inválidos fornecidos"
          }
        },
        "description": "Estrutura padrão de resposta para erros da API"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}